refactor: 프로젝트 구조 개선
Lotto.java, LottoTest.java 파일을 domain 패키지로 이동함.
message/ViewMessage.java 파일 삭제함.(domain/message 패키지에 존재)
