+++ src/main/java/lotto/domain/lotto/Lotto.java
@@ -1,0 +1,27 @@
+package lotto.domain.lotto;
+
+import java.util.List;
+import lotto.domain.message.ErrorMessage;
+
+// 필드 추가 금지!!
+public class Lotto {
+    private final List<Integer> numbers; // private 변경 금지
+
+    public Lotto(List<Integer> numbers) {
+        validate(numbers);
+        this.numbers = numbers;
+    }
+
+    private void validate(List<Integer> numbers) {
+        if (numbers.size() != 6) {
+            throw new IllegalArgumentException(ErrorMessage.INVALID_NUMBER_COUNT.getMessage());
+        }
+
+        numbers.stream()
+                .filter(number -> number < 1 || number > 45)
+                .findAny()
+                .orElseThrow(() -> new IllegalArgumentException(ErrorMessage.INVALID_NUMBER_RANGE.getMessage()));
+    }
+
+    // TODO: 추가 기능 구현
+}
