+++ src/test/java/lotto/domain/lotto/LottoTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain.lotto;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
